Product supplier endpoint refactor

+++ src/main/java/com/utn/ProgIII/dto/CreateProductSupplierDTO.java
@@ -16,5 +16,5 @@ public record CreateProductSupplierDTO(
         @Schema(example = "250")
         @PositiveOrZero(message = "El costo no puede ser menor a cero")
         @NotNull(message = "El costo no puede ser nulo")
-        BigDecimal cost) {
+        Double cost) {
 }
